Changes, entity-aware API is complete
