+++ trustgraph-base/trustgraph/knowledge/__init__.py
@@ -1,4 +1,5 @@
 
+from . defs import *
 from . identifier import *
 from . publication import *
 from . document import *
